add text and metadata extraction to ETL
